requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.2
